requirements.txt
Django==5.1.2
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
mysqlclient==2.2.5